Implement new status page
This is a rework of the current status change with focus on a more
operator-oriented view. The idea is to show the current status of the
tenant ("What's currently going on") via a minimalistic approach.

The new page doesn't show all information about the pipelines like
before, which is by purpose to keep the focus on a higher level
(overview of a single tenant).

The next change will implement a PipelineDetails page that will
compensate those missing information and also show more details about a
single pipeline.

Some of the key-features are:

  1. Better overview on the overall state of a tenant, with less details
     on individual queues and items. This is especially useful for
     operators.
  2. Queues in dependent pipelines are more prominent. In the old status
     the queues weren't that visible and it was often not clear to which
     queue an item belonged.

The change doesn't touch the current StatusPage, but creates a new one
for easier review. The old status page will be cleaned up in a follow-up
commit.

Change-Id: I255e84ba562bb87da6bdbd44b51adbcc73136c47