Make helper functions available to other components
This moves some of the helper functions used to format and display
various information of a QueueItem to a more generic file so they can be
used by other components as well.

In addition, the render methods for the ChangeLink and
(QueueItem)Progressbar where changed to React Components and now use
newer PF4 components for styling.

Finally, this changes fixes a typo in the calculateQueueItemTimes
function to make the fallback to null work.

Change-Id: Ia7f3f214dfa1dbc041c1dcb821ff478980d81011