Implement PipelineDetails view
As an addition to the new status page, this implements a detail view for
a single pipeline. The pipelines details view can be reached by clicking
on the title of a pipeline on the new status page.

The information that is shown on this page is basically the same that
was shown on the old status page. However, by keeping the focus on a
single pipeline, the information can be presented in a more prominent
way using the whole screen/page. On the old status page the information
for each pipeline was very "squished" as each pipeline only got a little
bit of space.

Some key features:

  1. Queues in dependent pipelines are more prominent, so it's clear
     what is enqueued where.
  2. Non-live items are shown in a more compact way and also don't
     provide any actions (like dequeue and promote).

This change copies and adapts a few methods from the
containers/status/ItemPanel.jsx file to render job-specific information
(like progress, result and links). I didn't bother to "lift" those to
more generic files as this would also require some adaption to make them
"backward-compatible" (support the old and new styling). On the long run
the single change panel (ItemPanel.jsx) will also be replaced and we can
deal with that later.

TODOs:
  1. Branching: The "branching" feature is still missing, so items are
     not branched out in case of e.g. a MERGE_CONFLICT in a dependent
     pipeline. The idea would be to keep the branching a little simpler.
     Items that are no longer relevant (e.g. due to a merge conflict)
     are greyed out and explicitly disconnected from the active
     "thread".

Change-Id: I9737c80337d8156f8fc8fc521c540aa09d4a6a18