Implement PipelineDetails view
As an addition to the new status page, this implements a detail view for
a single pipeline. The pipelines details view can be reached by clicking
on the title of a pipeline on the new status page.

The information that is shown on this page is basically the same that
was shown on the old status page. However, by keeping the focus on a
single pipeline, the information can be presented in a more prominent
way using the whole screen/page. On the old status page the information
for each pipeline was very "squished" as each pipeline only got a little
bit of space.

Some key features:

  1. Queues in dependent pipelines are more prominent, so it's clear
     what is enqueued where.
  2. Non-live items are shown in a more compact way and also don't
     provide any actions (like dequeue and promote).

This change copies and adapts a few methods from the
containers/status/ItemPanel.jsx file to render job-specific information
(like progress, result and links). I didn't bother to "lift" those to
more generic files as this would also require some adaption to make them
"backward-compatible" (support the old and new styling). On the long run
the single change panel (ItemPanel.jsx) will also be replaced and we can
deal with that later.

Additionally, this aligns the layout and styling of the QueueItemPopover
to match the QueueItem. Where possible those parts are also moved to a
common file, so they are only defined once.

Please note that the "branching" feature is currently missing, so items
are not branched out in case of e.g. a MERGE_CONFLICT in a dependent
pipeline. This feature will be implemented in a follow-up change.

Change-Id: I9737c80337d8156f8fc8fc521c540aa09d4a6a18