Add filter actions menu to error table
Continuing the attempt to find a nice way to add filtering to
table rows: rather than put a filter icon next to each field, use
the patternfly react-table actions menu feature to put a single
search icon on the right of the table row which pops up a menu
that allows users to add a filter for each filterable column.

This gives us the full interactive filter selection functionality
without cluttering up the table.

The action menu is not compatible with our custom "zuul-table" css,
so that class has been removed here.  The main difference at this point
is that it doesn't have the animated highlight on the left side during
mouseover.

Change-Id: I3ebac23296e61cfafcec3d5b37e65fffa4baad40