Only use latest proposed config for project-branch
When an item updates config we will schedule a merge for the proposed
change and its dependencies.

The merger will return a list of config files for each merged change.
The scheduler upon receiving the merge result will combine the collected
config files for a project-branch from all involved changes.

This lead to the problem that the old content of renamed config files
were still used when building the dynamic layout.

Since the config we receive from the merger is always exhaustive, we
just need to keep the latest config files.

Another (or additional) fix would be to only return the latest config
files for a project-branch from the mergers. However, in case of
circular dependencies it could make sense in the future to get the
update config per change to report errors more precisely.

Change-Id: Iebf49a9efe193788199197bf7846e336d96edf19