Always delay response to first build request
In Ic52f3d82ed12fd45922d9ec9d61e8ded43507837 we changed the build
worker to delay answering request only after we've successfully locked
and started a job.

This change also implicitly removed the initial delay to the first build
request. However, in practice this turned out to be a mistake, as
low-latency executors where capable of responding much quicker to
request the others with a higher ZK client latency. Since during normal
load there are usually just a few build request in each iteration
reinforces the problem.

The effect is that builds are no longer equally distributed among
executors. To fix this problem we'll add back the initial delay before
responding to the first build request.

Change-Id: I91f2ce87ccc2c5b6a854cf7937ebe65521e72d66