Revise decision to not deduplicate noop jobs
Since circular dependencies are now modelled as multiple changes in a
single item there is no good reason anymore not to deduplicate noop
jobs.

So far the noop job was listed separately for each change that was part
of a cycle and with that cluttered the UI, especially for large
dependency cycles.

Change-Id: Ic8d447df7d9040a767c88127ba361badc9ef016a