[WIP] fix autohold for gitlab merge requests
when using the web UI to autohold future builds for a change hosted on gitlab,
zuul-web returns an error 500 because the gitlabsource driver doesn't know
how to convert a change into a ref.

According to various sources and if I read the gitlab API correctly,
the MR ref should be in the form "refs/merge-requests/:iid:/head", which
is already set as change.ref when instantiating a MR.

Fix getRefForChange to return this ref.

Change-Id: If600c629300158ae71f168c601f64ae99410744c