Update patch set 3
Patch Set 3: Code-Review+2

This is what you mentioned in Iebf49a9efe193788199197bf7846e336d96edf19 -- it's more efficient in that we are returning less data, but it may constrain us in the future so that we don't know which change in a cycle that a file came from, right?

I think I agree this is reasonable to merge for now, and we can revert it if we decide we need the extra resolution.

Patch-set: 3
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2, ffcaf2082c90760bfab51d09c4cc42aef9088cf8