Use the triggering change as the zuul change
In the case of circular dependencies with job deduplication, we
arbitrarily pick one of the changes as the zuul change (to use
when setting the zuul.change job variable and friends).  In
theory, it shouldn't matter which change we use, but in practice,
users might use this value to coordinate with external systems.
Since it doesn't really matter to Zuul, let's set the zuul change
to the triggering change when possible.  It still needs to be one
of the changes for the job, so if the triggering change itself
doesn't actually run the job (easily possible if the job is only
run on dependent changes), then we will fall back to the current
behavior.  And of course the change must be one of the items
changes, so in the case of linear dependencies, we're not going
to start setting it to some other queue item's change.

If we are unable to set it to the triggering change, then the
behavior remains undefined beyond setting it to one of the job's
changes arbitrarily.

Included in this change is a cleanup of a no-longer-needed api
migration from 12->13 related to EventInfo objects that was
missed due to a missing MODEL_API tag.

Information about the triggering change is added to the EventInfo
object to implement this feature.

Because the fallback behavior and the model upgrade behavior are
the same, we don't need to add any conditional api behavior or
upgrade testing -- in both cases we will simply use the current
behavior.

Change-Id: Iee5a7d975fea1f7491b652c406c24d73ada7a1a1