Allow a Streamer to point to a custom host
Console Stream instances, when created, points to the IP/host name
of fetched from ansible_host (or ansible_inventory_host if the
previous one is not present). In certain scenarios, where hosts
are added dynamically and in a network that is not directly
reachable by Zuul, we may want to point the Stream object for a
given host to a host, reachable from Zuul, that tunnels the socket
from the unreachable host to a Zuul instance.
In that scenario the user should take care of ensuring the console
socket is properly tunneled to a host that zuul can reach, and pass
the new couple of variables in the ansible.builtin.add_host call
to tell Zuul where to connect its socket.

Change-Id: I8e747e09265699cd7be612c15361340a2e4a8f40