Fix web dequeue of post items
This is a partial revert of 6c1c91f8ba6305f7c742133ca99bc0253ea6774f
which relied on change.id being absent in the case that a queue item
was for a non-change ref; however, all queue items have an id.  This
reverts that section of code back to the previous regex-based
heuristic.

Change-Id: I07ff7c2925b44a7e85529491755b2f3d3f73eaf8