Replace status_url with item_url in pipeline reporter templates
We have some replacement fields designed to allow users to construct
deep links to the status page, but that is more difficult to do now
that items may have more than one change.  Moreover, users shouldn't
need to do it at all since Zuul already knows how to do that.

Make a new field, "item_url" which is the correct status page url,
and encourage users to use that instead.  Deprecate status_url,
change, and changes which are now difficult to use correctly.

Change-Id: I40e98ed220a13b8d2edddf510bab133b05845751