Change allow_pre_fail phases if will_retry
Early failure detection causes unnecessary rebuilds in dependent
pipelines. If buildset A and buildset B are building in a dependent
pipeline and a build in buildset A fails in a pre-run step, which
then causes the post-run step to fail, all the jobs in buildset B will
restart because a failure was detected. Instead, we should only detect
failures in the post-run step if we are not going to retry.

Change-Id: Iccc70c59831c27fb8ee512d49e3899ee705b35d1