Use the pipeline from the buildset
There's no pipeline key in the ref object. This causes the queue
action to fail with a 400 error due to the missing pipeline key
in the request payload.

This change switches to using the buildset.pipeline value, which was
being used in the past and changed in 24e1cbc26d252259a562ddc186cdedf6f4aa7bea

Signed-off-by: Flavio Percoco <flavio@pacerevenue.com>
Change-Id: Icaf65d798b03c083944280140d25267351d61f78