Replace pkg_resource dep resolution with pip
Python 3.12 has more fully excised setuptools and as a result
pkg_resources. Since we are using pip to install things we can instead
rely on pip to resolve whether or not packages are installed. Do this so
that we don't need to explicitly install setuptools where it may not be
needed.

Change-Id: I8ee189c828914fd648847b5650b5fb2fb255ff17