Sort pipelines on status page by number of queue items
Currently, the pipelines are ordered in the way they are loaded by
Zuul during config loading. Depending on how the pipelines are
structured within the Zuul config files for each tenant, the ordering
might look very messy (in the Zuul UI). Most of the time it's neither
alphabetically nor by pipeline type.

To go in hand with the idea behind the new status page, this change
sorts the pipelines by the number of enqueued items. Pipelines that
have a lot of items enqueued should come first, while pipelines with no
items enqueued should come last.

In addition to that, the pipelines are also sorted by their pipeline
type, so that e.g. dependent pipelines come before inpendentent
pipelines.

The idea behind this to sort the pipelines by their "activity", which
fits well to the approach of the new status few and keeps the focus on
what's going on in the system while trying to reduce the distraction
(e.g. by shifting empty pipelines to the end of the page). This way,
one can clearly see what is running in the system and where.

Change-Id: I24ffa7cb0c21bebf67f05c3f436e13fe2509cf0b