Hide empty pipelines and queues on status page
To improve the focus of the new status page, filter out empty pipelines
and queues. This can be disabled via a switch (toggle).

Both pipelines and queues are controlled by the same toggle, but it
might make sense to add a dedicated toggle for each option.

By default empty pipelines and queues are still hidden as this makes the
new status page much more concise.

Change-Id: Iae182425cc7e5f489ac685d24d6ef077fb4fa05f