Align job progress bars and job result labels
This aligns the layout of the job progress bar and job result lables
which are shown in the job list (expandable section of a QueueItem).

The idea is to keep the styling similar while one should be able to
clearly distinguish between a running job (progress) and a completed job
(result).

Change-Id: I57c4cc1bb1156288fd7e2d14b458e3abec009a40