Store a repo state file in the log directory
This adds a file named workspace-repos.json to the workspace build
log directory.  It contains the repo state and git commands needed
to reconstruct the repo state used for the job.

A later change may add a tool to translate the structured data in
this file to a shell script.

Other uses of the file may involve reading the repo state directly
to determine the last "real" commit in the tree before Zuul started
creating its speculative vision of the future.

Change-Id: Ifc5ee03fdc41a82f481ac57b8992ba7bd5f9b6c0
Co-Authored-By: Tobias Henkel <tobias.henkel@bmw.de>