Update patch set 6
Patch Set 6:

I just realized that I have plans to introduce a new variable with the refs for the specific build in the case that a build was deduplicated.  To disambiguate that, I think we should call this "buildset_refs" instead of refs, and call the later variable "build_refs".  So we will have:

* zuul.buildset_refs: list of all refs for the item (the entire circular dependency set)
* zuul.build_refs: list of refs that run this job (the subset of the circular dep that caused this particular build to run in the case of deduplication)

Patch-set: 6