Add buildset event db table
This adds a new database table, zuul_buildset_event, which stores
event information related to the buildset.  At the moment, the
only event information stored is a description of the trigger event
which originally caused the item to be enqueued.  Later changes may
add other events such as the reason the item was dequeued or
superceded, or the reason a buildset was canceled due to a gate reset.

A particular goal is to be able to inform users who see unsolicited
reports on their changes (which can happen due to a different change
in a dependency cycle being enqueued) understand why the report was
made.

The new event information is exposed via the rest api and a later
change will add it to the web ui.

Change-Id: I9bcbb64faa3d499a26b90d20932009b6ce226061