Move fake gerrit and pagure into dedicated files
This change is merely a reorganization to move the fake gerrit and
gitlab classes into their own files to match github and gitlab.

The Fake*Connection classes for all 4 drivers are also moved into
their respective files.  This is accomplished by moving some symbols
from base.py into a new tests/util.py to resolve the import cycle
(which is likely why they were not there in the first place).

Change-Id: I274b9e5abf6086656f8ceb5a16dab2f8393deead