Make the test change database serializable
This adds methods to save and restore the test change database.
This can be used in a future change to perform an upgrade test
where we save and restore the fake gerrit (and other drivers)
state across an upgrade.

Some minor changes are made to FakeGerritChange to avoid storing
non-pickleable values.  Similarly, the Gerrit reporter was using
ZuulConfigKey objects as review categories.  That's fine since
they behave like strings everywhere we use them, but they aren't
pickleable, and we don't need the extra line context information
after loading the config, so we discard it and convert them to
strings.

Change-Id: Ifa404203e414932f50811291ad56a661f2875af0