Handle tags when packing refs
Zuul packs refs directly rather than rely on git to do so. The reason
for this is it greatly speeds up repo resetting. Typically there are two
pieces of information for each packged ref (sha and refname), but git
tags are special because they have the sha of the tag object project,
the tag refname, and finally the sha of the object the tag refers to.

Update Zuul's ref packing to handle this extra piece of information for
git tags.

Change-Id: I828ab924a918e3ded2cd64deadf8ad0b4726eb1e