wip: Prototype cooperative zuul-launcher
This change prototypes an alternative node request handling strategy to
what was proposed in the original spec[0] and tries to benchmark it
against the current Nodepool implementation and the spec proposal.

Instead of grabbing all requests a launcher can get - as it is the case
with the current Nodepool launchers - we'll use the component registry
and consistent hashing (rendezvous / highest random weight hashing) to
create a priority list of launchers. The launcher with the highest score
will lock and process the request.

When a new launcher shows up it won't process any locked requests, even
though it might have a higher score than existing launchers.

When a launcher disappears the request will be unlocked and the
remaining launchers will decide based on the score who is to process the
request.

The advantage of this approach is that we avoid thundering herd effects
when processing requests and launchers can be more efficient due to data
locality.

Open topic: how to handle provider quota calculation

[0] https://zuul-ci.org/docs/zuul/latest/developer/specs/nodepool-in-zuul.html

Change-Id: Icced34d9c8d9e2743a0644fad0e4d82665742e87