wip: Prototype cooperative zuul-launcher
This change prototypes an alternative node request handling strategy to
what was proposed in the original spec[0] and tries to benchmark it
against the spec proposal and the current Nodepool implementation.

Instead of grabbing all requests a launcher can get (as it is the case
with the current Nodepool launchers) or locking a provider/global
request queue (as proposed in [0]) we'll use the component registry and
consistent hashing (rendezvous / highest random weight hashing) to
create a priority list of launchers. The launcher with the highest score
will lock and process the request.

When a new launcher shows up it won't process any locked requests, even
though it might have a higher score than existing launchers.

When a launcher disappears the request will be unlocked and the
remaining launchers will decide based on the score who is to process the
request.

The advantage of this approach is that we avoid thundering herd effects
when processing requests (which we see in the Nodepool launcher today)
and launchers can be efficient due to data locality.

In addition we don't need any explicit coordination as proposed in [0],
with the goal to process requests - globally and on the provider level -
in parallel.

Due to that level of parallelization we need to also consider handling
of quota (provider and tenant) as well as provider rate limits.

The Nodepool launcher implementation as it is today will perform the
check whether there is any remaining quota independently for each node
request based on cached information about existing nodes. This means
that concurrently processed requests in different provider pools
(Nodepool only supports one launcher per provider) will not consider
each other's resource usage and there might be a small delay until new
nodes show up in the cache.

Same is true for the tenant quota that also considers resources used by
other providers, so there can be more requests in parallel that might
not consider each other's resource requirements.

With the cooperative ("implicitly coordinating") launcher approach
proposed here, the main difference is, that the possibility for quota
races would increase when scaling up the launcher instances (more
requests are processed in parallel). The accuracy of provider quota
calculations would probably be similar to that of the tenant quota
today.

Ways to account for that effect on calculated quota (in order of
preference):

* provider quota calculation is best-effort similar to the tenant quota
  today; consider/document potential "overshoot" when setting resource
  limits
* calculate quota when assigning requests to providers, but also
  re-calculate/check the quota limits on provider level and reject a
  request if it would exceed the quota limits (used quota might still be
  a bit off)
* use a combination of explicit coordination (lock) for the global
  request queue (inc. quota calculation) as proposed in the spec and
  implicit coordination for provider processing as proposed in this
  change (this might allow more precise quota calculations than what we
  have today)
* synchronize quota reservation/calculation (optional?; might slow down
  overall request processing)

Rate limiting in Nodepool today works based on a rate-limiter with the
rate configured at the provider level. Multiple provider pools will
respect the global provider rate limit. With multiple launchers for a
single provider (as proposed in this change) the limit needs to be
shared between the launchers. This could be done statically via the
config or based on the online launchers with information from the
component registry.

Another option would be to adjust the rate limit more dynamically based
on response headers and/or rate limit errors, but that might not be
worth the additional complexity.

[0] https://zuul-ci.org/docs/zuul/latest/developer/specs/nodepool-in-zuul.html

Change-Id: Icced34d9c8d9e2743a0644fad0e4d82665742e87